# Fixed the animation of main bar. Finally, our main page animation is working.
## Previous Implementation
The previous implementation involved having two separate components one for the main page and one for the sidebar but in Non-Mobile views. What we basically did was move this outer container from one layout to another. But the main problem we were facing was the layout shift that was caused by the removal of upper section of that component.

## Current Implementation *working*
Our current implementation improves the layout transition when we click on some route like new pipeline or inference and moving to a non-mobile view. So what we are basically doing here, we are deriving the state of the component from the same component. So it's now generating two different views, one for the home route and one for the new pipeline and inference route but for non-mobile views only because on mobile we have a separate sidebar that we developed in previous comments.

# Different Logo for diff modes
So, up till now our application was generating the same main logo for both light and dark mode. And what's happening now is that we are deriving based on our state of the current theme that we should have a different logo for different theme like for the dark and light.

## Tailwind Theme bug
So we were facing a weird bug that we were unable to derive the light and dark clause from the CSS. We are saving it as a state also, but it should be really convenient for us to use that the same theme but from our style files.

The same was fixed by deriving the state of our theme from class of dark on our root i.e. `<html>` and using the following in main `.css` file
```css
@custom-variant dark (&:where(.dark, .dark *));
```
So what basically does it change the default behavior to derive the theme just from the system directly to move a level down and derive it from the state of what we are setting as a class in HTML when the user toggles the theme.