# Schema overhauled This schema was written again and the defaults now lie in the helper function of pipeline dl. So, any default should absolutely be changed from there and it is the only source of truth for defaults and schema folder now serves as only source of truth for pipeline dl schema.
# Pre-processing tab updated.
Pre-processing tab for image processing, audio processing and text processing were added as a main task.

# Transformer Param update
A very intense feature was added to the pre-processing tab that apart from each tab being reorderable each one has a gear icon which when triggered will show a dialog containing the form to change the parameters of that particular transform.

# Transform params updating strategy re-engineered
Updating transform params now does not depend on the main task.
This comes from the fact that transformer names cannot overlap across main tasks as mentioned by @haroon.
So, if they overlap, for example, like ToTensor, then ToTensor will or has to share the same schema across all the mainTask it falls in. (which as of now it does)

# Command added
Command component was added to give the functionality of searching the pre-processing.

# InputNumberArray and InputStringArray components were formed.
These both components lay down the foundation for us to get the input as array of numbers and array of strings from the user.

> All the major changes in this comment have been documented; for any further information, you may look at the git log.